Rasptimer Solar Timer Scheduler
This is based on the excellent original concept and program by Johannes
Ernst for his pool timer and then some later additions by Alan Stead. It
also makes use oif the very useful Wiring Pi software by Gordon
Henderson.

Essentially the Rasptimer is a web based timer such as you may have
plugged into your AC wall socket except that it can manage any number of
devices/appliances within reason (currently 6) with a number of
different schedules and of course it has greater ease of use.

This version currently uses Johannes original Raspberry Pi Raspbian
Apache server version and has a number of additional features to allow
the timer to be used with solar PV installations. The software features
include a day-of-week programming, suspend, cloud (with additional low
light detector), bump or nudge feature for the timers and more
importantly 4 timer auto-schedules per appliance. Thus power usage may
be carefully controlled to make greater use of the solar power
available.

There is also a new hardware/software feature to make the timer more
user friendly for anybody who does not have solar and this is the
inclusion of WiFi AC sockets. This is currenly under delevopment but is
simply a $15 off-the-shelf WiFi enabled AC socket similar to a mobile
double socket you probably already have except that it is switch driven
by the Solar scheduler via a WiFi link and it can be plugged in anywhere
within about a 30m range.
A temperature module is also going to be added

The rest of the hardware information will be included in another
document.